editing post enabled

+++ app/Http/Livewire/Connect/Post/EditPost.php
@@ -4,39 +4,62 @@ namespace App\Http\Livewire\Connect\Post;
 
 use App\Http\Livewire\Traits\CreatesSocialContent;
 use Livewire\Component;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\Storage;
 
 class EditPost extends Component
 {
-    use CreatesSocialContent;
+    use CreatesSocialContent,
+    AuthorizesRequests;
 
     public $post;
     public $gallery;
+    public $confirm = false;
     protected $listeners = [
         'refreshMe' => '$refresh'
     ];
 
-    public function mount()
-    {
-        $this->text_content = $this->post->content;
+    public function mount() {
+        $this->authorize('update', [$this->post, auth()->user()->currentProfile]);
+        $this->text_content = (string) $this->post->content;
         if ($this->post->gallery->count() > 0) {
             $this->hasStoredImages = true;
             $this->gallery = $this->post->gallery;
         }
     }
 
-    public function create()
-    {
-        $this->validate($this->validationRules());
+    public function confirmDeletePost() {
+        $this->confirm = true;
     }
 
-    public function removeFromStoredPhotos($image)
-    {
-        $key = $image;
+    public function create() {
+        $this->validate($this->validationRules());
+        $this->post->content = $this->text_content;
+        $this->post->save();
+        if (count($this->photos) > 0) {
+            $this->uploadPhotos('post-photos', $this->post, 'post_photo');
+            $this->photos = [];
+            return $this->redirect($this->post->url->edit);
+        }
+        return $this->emitSelf('saved');
+    }
+
+    public function deletePost() {
+        $this->authorize('update', [$this->post, auth()->user()->currentProfile]);
+        $this->post->trash();
+        return $this->redirect(route('home'));
+    }
+
+    public function removeFromStoredPhotos($image) {
         $image = $this->gallery->find($image);
-        Storage::disk('public')->delete($image->image_url);
-        $image->delete();
-        return;
+        if ($this->post->gallery->count() > 1 || $this->post->content !== '') {
+            $image_url = $image->image_url;
+            $image->delete();
+            $this->emitSelf('refreshMe');
+            return Storage::disk('public')->delete($image_url);
+        } else {
+            $this->addError('last_content', 'deleting will empty post content');
+        }
     }
 
     public function extra_validation(): array
@@ -44,8 +67,8 @@ class EditPost extends Component
         return [];
     }
 
-    public function render()
-    {
+
+    public function render() {
         return view('livewire.connect.post.edit-post');
     }
 }
